refactor: move SourceType from oxc_ast to oxc_span
related #350
