+++ crates/oxc_cli/Cargo.toml
@@ -19,10 +19,10 @@ mimalloc = { workspace = true }
 [dependencies]
 oxc_diagnostics = { workspace = true }
 oxc_allocator   = { workspace = true }
-oxc_ast         = { workspace = true }
 oxc_parser      = { workspace = true }
 oxc_semantic    = { workspace = true }
 oxc_linter      = { workspace = true }
+oxc_span        = { workspace = true }
 
 clap       = { workspace = true }
 rayon      = { workspace = true }
