+++ crates/oxc_formatter/Cargo.toml
@@ -12,6 +12,7 @@ repository.workspace  = true
 [dependencies]
 oxc_allocator = { workspace = true }
 oxc_ast       = { workspace = true }
+oxc_span      = { workspace = true }
 oxc_syntax    = { workspace = true }
 
 [dev_dependencies]
