+++ crates/oxc_minifier/src/lib.rs
@@ -6,9 +6,9 @@ mod compressor;
 mod printer;
 
 use oxc_allocator::Allocator;
-use oxc_ast::SourceType;
 use oxc_ast_lower::AstLower;
 use oxc_parser::Parser;
+use oxc_span::SourceType;
 
 pub use crate::compressor::CompressOptions;
 use crate::compressor::Compressor;
