+++ crates/oxc_parser/examples/parser.rs
@@ -1,8 +1,8 @@
 use std::{env, path::Path};
 
 use oxc_allocator::Allocator;
-use oxc_ast::SourceType;
 use oxc_parser::Parser;
+use oxc_span::SourceType;
 
 // Instruction:
 // create a `test.js`,
