+++ crates/oxc_span/src/source_type.rs
