+++ crates/oxc_wasm/Cargo.toml
@@ -15,12 +15,12 @@ default = ["console_error_panic_hook"]
 
 [dependencies]
 oxc_allocator   = { workspace = true }
-oxc_ast         = { workspace = true, features = ["serde"] }
 oxc_diagnostics = { workspace = true }
 oxc_parser      = { workspace = true }
 oxc_semantic    = { workspace = true }
 oxc_linter      = { workspace = true }
 oxc_formatter   = { workspace = true }
+oxc_span        = { workspace = true }
 
 miette = { workspace = true, features = ["fancy-no-backtrace"] }
 serde  = { workspace = true }
